# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NA Notes API
  description: >
    Shared REST contract implemented identically by every backend-* folder
    in this repository, and consumed identically by every frontend-*
    folder. See /docs/adr for the reasoning behind the auth, pagination
    and concurrency design used here.
  version: 1.0.0
servers:
  - url: /api
    description: Relative to whichever origin the backend is deployed on.
...
security:
  - sessionCookie: []
paths:
...
  /notes/{noteId}/public-share:
    parameters:
      - name: noteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [sharing]
      operationId: createPublicShare
      summary: Publish a note publicly (read-only), owner only
      description: Generates a new unguessable share token (ADR 0009).
      responses:
        "201":
          description: Public share created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicShare"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [sharing]
      operationId: deletePublicShare
      summary: Revoke public sharing, owner only
      description: Invalidates the current public share token; the note is no longer publicly readable.
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: >
        Opaque server-side session identifier, HttpOnly + Secure +
        SameSite=Lax (ADR 0004). State-changing requests additionally
        require the csrfToken security requirement below.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >
        Must match the value of the non-HttpOnly `csrf_token` cookie
        (ADR 0005). Required on POST/PUT/PATCH/DELETE.
  responses:
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Logged in, but not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code, e.g. VERSION_CONFLICT.
            message:
              type: string
...
    PublicShare:
      type: object
      required: [token, url, createdAt]
      properties:
        token:
          type: string
        url:
          type: string
          format: uri
          description: Fully qualified public view URL for the frontend to display.
        createdAt:
          type: string
          format: date-time
...